+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # `stats_arrays` Changelog
 
+# 0.6 (dev)
+
+* Allow `MCRandomNumberGenerator` to generate multiple samples at once
+* Move tests to pytest
+* Add tests for MCRNG
+
 ## 0.5.1 (11-05-2019)
 
 * Import from `collections.abc` when possible, for Python 3.8 support.